Moved notify gatekeeper about to submit method to new controller
